Move zoom methods into a hook.
